Move events into client and server submodules
And use plural for plugin names.
