+++ src/client/events.rs
@@ -1,4 +1,4 @@
-mod client_event_data;
+mod event_data;
 
 use std::io::Cursor;
 
@@ -9,8 +9,8 @@ use bevy::{
 use bincode::{DefaultOptions, Options};
 use serde::{de::DeserializeOwned, Serialize};
 
+use super::{replicon_client::RepliconClient, server_entity_map::ServerEntityMap, ClientSet};
 use crate::{
-    client::{replicon_client::RepliconClient, server_entity_map::ServerEntityMap, ClientSet},
     core::{
         channels::{RepliconChannel, RepliconChannels},
         common_conditions::*,
@@ -19,7 +19,7 @@ use crate::{
     },
     server::{replicon_server::RepliconServer, ServerSet},
 };
-use client_event_data::ClientEventData;
+use event_data::ClientEventData;
 
 /// An extension trait for [`App`] for creating client events.
 pub trait ClientEventAppExt {
@@ -148,9 +148,9 @@ impl ClientEventAppExt for App {
     }
 }
 
-pub struct ClientEventPlugin;
+pub struct ClientEventsPlugin;
 
-impl Plugin for ClientEventPlugin {
+impl Plugin for ClientEventsPlugin {
     fn build(&self, app: &mut App) {
         app.init_resource::<ClientEventRegistry>()
             .add_systems(
@@ -174,7 +174,7 @@ impl Plugin for ClientEventPlugin {
     }
 }
 
-impl ClientEventPlugin {
+impl ClientEventsPlugin {
     fn send(world: &mut World) {
         world.resource_scope(|world, mut client: Mut<RepliconClient>| {
             world.resource_scope(|world, registry: Mut<AppTypeRegistry>| {
