+++ src/server/events.rs
@@ -1,4 +1,4 @@
-mod server_event_data;
+mod event_data;
 
 use std::io::Cursor;
 
@@ -7,6 +7,9 @@ use bincode::{DefaultOptions, Options};
 use ordered_multimap::ListOrderedMultimap;
 use serde::{de::DeserializeOwned, Serialize};
 
+use super::{
+    connected_clients::ConnectedClients, replicon_server::RepliconServer, ServerPlugin, ServerSet,
+};
 use crate::{
     client::{
         replicon_client::RepliconClient, server_entity_map::ServerEntityMap, ClientPlugin,
@@ -19,12 +22,8 @@ use crate::{
         replicon_tick::RepliconTick,
         ClientId,
     },
-    server::{
-        connected_clients::ConnectedClients, replicon_server::RepliconServer, ServerPlugin,
-        ServerSet,
-    },
 };
-use server_event_data::ServerEventData;
+use event_data::ServerEventData;
 
 /// An extension trait for [`App`] for creating client events.
 pub trait ServerEventAppExt {
@@ -151,9 +150,9 @@ impl ServerEventAppExt for App {
     }
 }
 
-pub struct ServerEventPlugin;
+pub struct ServerEventsPlugin;
 
-impl Plugin for ServerEventPlugin {
+impl Plugin for ServerEventsPlugin {
     fn build(&self, app: &mut App) {
         app.init_resource::<ServerEventRegistry>()
             .add_systems(
@@ -179,7 +178,7 @@ impl Plugin for ServerEventPlugin {
     }
 }
 
-impl ServerEventPlugin {
+impl ServerEventsPlugin {
     fn send(world: &mut World) {
         world.resource_scope(|world, mut server: Mut<RepliconServer>| {
             world.resource_scope(|world, registry: Mut<AppTypeRegistry>| {
